♻️ Consolidate queues, add dataset status lookup, fix prometheus
