+++ aleph/views/entities_api.py
@@ -21,8 +21,7 @@ from aleph.logic.html import sanitize_html
 from aleph.logic.profiles import pairwise_judgements
 from aleph.model.bookmark import Bookmark
 from aleph.model.entityset import EntitySet, Judgement
-from aleph.procrastinate.queues import queue_export_search
-from aleph.queues import OP_EXPORT_SEARCH
+from aleph.procrastinate.queues import OP_EXPORT_SEARCH, queue_export_search
 from aleph.search import (
     DatabaseQueryResult,
     EntitiesQuery,
